mas bum bum

+++ trunk/src/Zona.cpp
@@ -99,7 +99,7 @@ Zona::~Zona() {
  */
 Zona* Zona::divAccion() {
 	
-	int x3 = (this->get_x2() + this->get_x1())/2;
+	int x3 = (int)(this->get_x2() + this->get_x1())/2;
 		
 	//creo una nueva Zona
 	Zona *zn = new Zona(x3+1,this->get_y1(),this->get_z1(),this->get_x2(),this->get_y2(),this->get_z2());
